% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bb_wls_system_cpp}
\alias{bb_wls_system_cpp}
\title{Weighted least-squares cross-products for beta-binomial IRLS}
\usage{
bb_wls_system_cpp(x, weight, response)
}
\arguments{
\item{x}{Numeric design matrix.}

\item{weight}{Positive working-weight vector.}

\item{response}{Numeric working-response vector.}
}
\value{
A list with \code{information} and \code{score_target}.
}
\description{
Computes the symmetric information matrix \eqn{X^\top W X} and score
target \eqn{X^\top W z} in compiled code. This is an internal kernel used
by \code{bbreg()}.
}
\keyword{internal}
